backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.9.2
numpy==2.1.3
scipy==1.14.1
pykrige==1.7.2
alembic==1.14.0
python-multipart==0.0.12
Pillow==11.0.0
